+++ CHANGELOG.md
@@ -6,6 +6,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Changed
+- Add conflict with origin neomerx/json-api (use fork laravel-json-api/neomerx-json-api for best php 8.* support)
+
 ## [0.0.15] - 2025-02-06
 
 ### Added